drum_transcription/pyproject.toml
[project]
name = "drum-transcription"
version = "0.1.0"
description = "Drum transcription model for rhythm game using E-GMD dataset"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
    "lightning>=2.4.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "mido>=1.3.2",
    "pretty-midi>=0.2.10",
    "numpy>=1.26.0,<2.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
    "h5py>=3.10.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "tensorboard>=2.15.0",
    "optuna>=4.7.0",
    "optuna-integration>=4.7.0",
    "hydra-core>=1.3.2",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.18.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# PyTorch CUDA configuration for Linux (training machines)
[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux'" }
]
torchvision = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux'" }
]
torchaudio = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux'" }
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
